Coredump, memory leak and missing container logs fix
This pull request addresses a number of issues \n Coredumps:\n
1. Modified  static vector<Container_DaemonEvent_Class> QueryAll() to only try to link stdout/stderr streams for newly created containers. It looks for create in status for all containers events gotten from docker dameon and then considers it a newly created container. This became a bug because it doesnt check whether the container is running now. Any container would go through a sequence of create/attach/stop/destroy. Its necessary to check if its currently running before trying to do anything with it\n
2. Modified static Container_ImageInventory_Class DeserializeObject(string& id). There were two issues here. A fixed buffer size was beign used. Since these files held a container inventory information I could see file sizes on my machine greater than 2KB. Its not unimaginable that someone might have file sizes greater than 4KB and so I decided to do away with the fixed file size and instead allocate a buffer based on the size of it. The other issue was that there was no check to see if it was able to suucessfuly parse that data. If it failed it would still try to access members and crash. So added that check\n
Memory leak : There was a bunch of places where memory was being allocated but never freed and so I tracked them down and added appropirate deallocation. A classic example was fprintf(target, "%s", cJSON_PrintUnformatted(root)); This can allocate upto 2KB of memory and the string is written into a file and ther is no one to deallocate it. Any free () or  cJSON_Delete() calls added were done to address this \n
Missing ContainerLogs: This had two issues. The first one was if there are identical logs only one of them would show up in OMS depsite all being posted to ODS. To make each record unique I added query docker logs with timestamps so each record becomes unique. The second issue was there was a 1 second timeout for the rest api call which is not enough in some cases. So the call would timeout, and the code would move forward asssuming there was no data. To resolve this I added a sort of a retrial mechanims with an increasing timeout. It would make six tries at max every time multiplying the timeout by a factor of 2. THe max timeout would be 64 seconds ~ 1minute. The third issue was there is some latency between an application writing to stdout/stderr and it showing up in docker logs. I observed this to be around ~ 3minutes. I increased the log upload frequencey to 5 minutes and that took care of this. This makes sure that there is definitely some logs which if generated were processed by docker. The fourht issue was the way the docker -provider handles the logs themseves. It uses a vector and pushes each record onto the end .Docker gives us the logs in a chronologically ascending order. Because the agent posts an entire batch at a time they all get the same TimeGenerated stamp. Since they all have the same timestamp the records appear in the order they were in the batch, earlier ones first. In order to get around this I used a deque instead and I add each record at the front. Switched to a double ended queue since its more efficient to add to the front compared to a vector.